+++ src/main/java/io/lenar/easy/log/JoinPointLogger.java
@@ -1,0 +1,112 @@
+package io.lenar.easy.log;
+
+import com.google.gson.Gson;
+import com.google.gson.GsonBuilder;
+import io.lenar.easy.log.annotations.LogCall;
+import io.lenar.easy.log.annotations.LogCalls;
+import io.lenar.easy.log.annotations.LogMethod;
+import io.lenar.easy.log.annotations.LogMethods;
+import io.lenar.easy.log.support.JoinPointSupport;
+import org.aspectj.lang.ProceedingJoinPoint;
+import org.joda.time.DateTime;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.lang.reflect.Method;
+import java.util.Map;
+
+public class JoinPointLogger extends JoinPointSupport {
+
+    private Gson gson = new GsonBuilder().serializeNulls().setPrettyPrinting().create();
+
+    protected static Logger logger = LoggerFactory.getLogger("EasyLogger");
+
+    public Object logCallMethodLevel(ProceedingJoinPoint jp, LogCall annotation) throws Throwable {
+        Method method = getMethod(jp);
+
+        logRequest(annotation.name(), method.getName(), getMethodParameters(jp));
+
+        int startTime = DateTime.now().getMillisOfDay();
+        Object result = jp.proceed(jp.getArgs());
+        int endTime = DateTime.now().getMillisOfDay();
+
+        logResponse(endTime - startTime, method.getName(), result);
+
+        return result;
+    }
+
+    public Object logCallClassLevel(ProceedingJoinPoint jp, LogCalls annotation) throws Throwable {
+        Method method = getMethod(jp);
+
+        logRequest(annotation.name(), method.getName(), getMethodParameters(jp));
+
+        int startTime = DateTime.now().getMillisOfDay();
+        Object result = jp.proceed(jp.getArgs());
+        int endTime = DateTime.now().getMillisOfDay();
+
+        logResponse(endTime - startTime, method.getName(), result);
+
+        return result;
+    }
+
+    public Object logMethodMethodLevel(ProceedingJoinPoint jp, LogMethod annotation) throws Throwable {
+        String methodName = jp.getSignature().getName();
+        Map<String, Object> params = getMethodParameters(jp);
+
+        if (params.isEmpty()) {
+            logger.info("-> Method " + jp.getSignature().getName() + " with no parameters");
+        } else {
+            logger.info("-> Method " + jp.getSignature().getName() + " with parameters \n" + gson.toJson(params) + "\n");
+        }
+
+        int startTime = DateTime.now().getMillisOfDay();
+        Object result = jp.proceed(jp.getArgs());
+
+        int endTime = DateTime.now().getMillisOfDay();
+
+        logger.info("Execution time: " + (endTime - startTime) + "ms");
+        logger.info("<- " + methodName + " returned: \n" + gson.toJson(result) + "\n");
+
+        return result;
+    }
+
+    public Object logMethodClassLevel(ProceedingJoinPoint jp, LogMethods annotation) throws Throwable {
+        String methodName = jp.getSignature().getName();
+        Map<String, Object> params = getMethodParameters(jp);
+
+        if (params.isEmpty()) {
+            logger.info("-> Method " + jp.getSignature().getName() + " with no parameters");
+        } else {
+            logger.info("-> Method " + jp.getSignature().getName() + " with parameters \n" + gson.toJson(params) + "\n");
+        }
+
+        int startTime = DateTime.now().getMillisOfDay();
+        Object result = jp.proceed(jp.getArgs());
+
+        int endTime = DateTime.now().getMillisOfDay();
+
+        logger.info("Execution time: " + (endTime - startTime) + "ms");
+        logger.info("<- " + methodName + " returned: \n" + gson.toJson(result) + "\n");
+
+        return result;
+    }
+
+    private void logRequest(String serviceName, String methodName, Map<String, Object> request) {
+        logger.info(requestMessage(serviceName, methodName, request));
+    }
+
+    private String requestMessage(String serviceName, String methodName, Map<String, Object> request) {
+        String message = "\n";
+        if (!serviceName.isEmpty()) message = message + serviceName + " CALL:\n";
+        return message + "-> " + methodName + " Request \n" + gson.toJson(request) + "\n";
+    }
+
+    private void logResponse(int responseTime, String methodName, Object result) {
+        logger.info(responseMessage(responseTime, methodName, result));
+    }
+
+    private String responseMessage(int responseTime, String methodName, Object result) {
+        return "Response time: " + responseTime + "ms\n" + "<- " + methodName + " Response: \n" + gson.toJson(result) + "\n";
+    }
+
+}
