+++ src/main/java/io/lenar/easy/log/annotations/LogCall.java
@@ -1,0 +1,14 @@
+package io.lenar.easy.log.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.METHOD)
+public @interface LogCall {
+
+    String name() default "";
+
+}
