+++ src/main/java/io/lenar/easy/log/annotations/LogMethod.java
@@ -1,4 +1,4 @@
-package io.lenar.easy.log;
+package io.lenar.easy.log.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,8 +7,6 @@ import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
-public @interface LogCall {
-
-    String name() default "";
+public @interface LogMethod {
 
 }
