+++ src/main/java/io/lenar/easy/log/annotations/LogMethods.java
@@ -1,4 +1,4 @@
-package io.lenar.easy.log;
+package io.lenar.easy.log.annotations;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -6,7 +6,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
-public @interface LogMethod {
+@Target(ElementType.TYPE)
+public @interface LogMethods {
 
 }
